requirements.txt
requests
dnspython
cryptography
